Struct bevy_tilemap::dimension::Dimension2

pub struct Dimension2 {
    pub width: u32,
    pub height: u32,
}

Dimensions of the 2nd kind.

Fields
width: u32

The width of this dimension.

height: u32

The height of this dimension.

Implementations

impl Dimension2

pub const fn new(width: u32, height: u32) -> Dimension2

Constructs a new 2nd dimension.

pub fn area(&self) -> u32

The total area of the dimension.

pub fn x_min(&self) -> i32

The minimum X value of this dimension.

pub fn y_min(&self) -> i32

The minimum Y value of this dimension.

pub fn x_max(&self) -> i32

The maximum X value of this dimension.

pub fn y_max(&self) -> i32

The maximum Y value of this dimension.

pub fn center(&self) -> Point2

Returns the center of the Map as a Vec2 Chunk coordinate.

pub fn check_point(&self, point: Point2) -> Result<(), DimensionError>

Checks if a coordinate is valid and inbounds.

Errors

If the point does not exist in the dimensions, an error is returned.

pub fn check_index(&self, index: usize) -> Result<(), DimensionError>

Checks if an index is valid and inbounds.

Errors

If the point does not exist in the dimensions, an error is returned.

pub fn encode_point_unchecked(&self, point: Point2) -> usize

Encodes a coordinate and returns an index value, unchecked.

pub fn encode_point(&self, point: Point2) -> Result<usize, DimensionError>

Encodes a coordinate and returns an index value.

Errors

If the point does not exist in the dimensions, an error is returned.

pub fn decode_point_unchecked(&self, index: usize) -> Point2

Decodes an index value and returns a coordinate, unchecked.

pub fn decode_point(&self, index: usize) -> Result<Point2, DimensionError>

Decodes an index value and returns a coordinate.

Errors

If the point does not exist in the dimensions, an error is returned.