Module bevy_tilemap::tilemap

Constructing a basic tilemap, setting tiles, and spawning.

Bevy Tilemap makes it easy to quickly implement a tilemap if you are in a rush or want to build a conceptual game.

use bevy_asset::{prelude::*, HandleId};
use bevy_sprite::prelude::*;
use bevy_tilemap::prelude::*;

// This must be set in Asset<TextureAtlas>.
let texture_atlas_handle = Handle::weak(HandleId::random::<TextureAtlas>());

let mut tilemap = Tilemap::new(texture_atlas_handle, 32, 32);

// Coordinate point with Z order.
let point = (16, 16);
let sprite_index = 0;
let tile = Tile { point: point.clone(), sprite_index, ..Default::default() };
tilemap.insert_tile(tile);

tilemap.spawn_chunk_containing_point(point);

Constructing a more advanced tilemap.

For most cases, it is preferable to construct a tilemap with explicit parameters. For that you would use a TilemapBuilder.

use bevy_asset::{prelude::*, HandleId};
use bevy_sprite::prelude::*;
use bevy_tilemap::prelude::*;

// This must be set in Asset<TextureAtlas>.
let texture_atlas_handle = Handle::weak(HandleId::random::<TextureAtlas>());

let mut tilemap = TilemapBuilder::new()
    .texture_atlas(texture_atlas_handle)
    .chunk_dimensions(64, 64, 1)
    .texture_dimensions(8, 8)
    .dimensions(32, 32)
    .add_layer(TilemapLayer { kind: LayerKind::Dense, ..Default::default() }, 0)
    .add_layer(TilemapLayer { kind: LayerKind::Sparse, ..Default::default() }, 1)
    .add_layer(TilemapLayer { kind: LayerKind::Sparse, ..Default::default() }, 2)
    .z_layers(3)
    .finish();

The above example outlines all the current possible builder methods. What is neat is that if more layers are accidentally set than z_layer set, it will use the layer length instead. Much more features are planned including automated systems that will enhance the tilemap further.

Setting tiles

There are two methods to set tiles in the tilemap. The first is single tiles at a time which is acceptable for tiny updates such as moving around characters. The second being bulk setting many tiles at once.

If you expect to move multiple tiles a frame, always use insert_tiles. A single event is created with all tiles if set this way.

use bevy_asset::{prelude::*, HandleId};
use bevy_sprite::prelude::*;
use bevy_tilemap::prelude::*;

// This must be set in Asset<TextureAtlas>.
let texture_atlas_handle = Handle::weak(HandleId::random::<TextureAtlas>());

let mut tilemap = Tilemap::new(texture_atlas_handle, 32, 32);

// Prefer this
let mut tiles = Vec::new();
for y in 0..31 {
    for x in 0..31 {
        tiles.push(Tile { point: (x, y), ..Default::default() });
    }
}

tilemap.insert_tiles(tiles);

// Over this...
for y in 0..31 {
    for x in 0..31 {
        tilemap.insert_tile(Tile { point: (x, y), ..Default::default() });
    }
}

Structs

Tilemap
A Tilemap which maintains chunks and its tiles within.

TilemapBuilder
Tilemap factory, which can be used to construct and configure new tilemaps.

TilemapError
The error type for operations when interacting with the tilemap.

TilemapLayer
A layer configuration for a tilemap.

Enums

ErrorKind
The kinds of errors that can occur.

Type Definitions

TilemapResult
A map result.