Struct bevy_tilemap::tilemap::TilemapBuilder

pub struct TilemapBuilder { /* fields omitted */ }

Tilemap factory, which can be used to construct and configure new tilemaps.

Methods can be chained in order to configure it. The texture_atlas method is required in order to have a successful factory creation.

The configuration options available are:

topology: sets the topology of the tilemap.

dimensions: specifies the dimensions of the tilemap. If this is not set, then the tilemap will have no dimensions.

chunk_dimensions: specifies the chunk's dimensions in tiles. Default is 32x, 32y.

texture_dimensions: specifies the tile's dimensions in pixels. Default is 32px, 32px.

layer_offset: Sets the layer offset as X, Y.

z_layers: specifies the maximum number of layers that sprites can exist on. Default is 20.

texture_atlas: specifies the texture atlas handle to use for the tilemap.

tile_scale: sets the tile scale in pixels.

add_layer: adds a layer to the tilemap.

auto_chunk: set if you want the tilemap to automatically spawn new chunks.

auto_spawn: set if you want the tilemap to automatically spawn and despawn chunks.

The finish method will take ownership and consume the builder returning a TilemapResult with either an TilemapError or the tilemap.

Examples

use bevy_asset::{prelude::*, HandleId};
use bevy_sprite::prelude::*;
use bevy_tilemap::prelude::*;

let texture_atlas_handle = Handle::weak(HandleId::random::<TextureAtlas>());

let builder = TilemapBuilder::new().texture_dimensions(32, 32).texture_atlas(texture_atlas_handle);

let tilemap = builder.finish().unwrap();

Can also get a builder like this:

use bevy_asset::{prelude::*, HandleId};
use bevy_sprite::prelude::*;
use bevy_tilemap::prelude::*;

let texture_atlas_handle = Handle::weak(HandleId::random::<TextureAtlas>());

let builder = Tilemap::builder().texture_dimensions(32, 32).texture_atlas(texture_atlas_handle);

let tilemap = builder.finish().unwrap();

Implementations

impl TilemapBuilder

pub fn new() -> TilemapBuilder

Configures the builder with the default settings.

Is equivalent to default and builder method in the tilemap. Start with this then you are able to method chain.

Examples

use bevy_tilemap::prelude::*;
use bevy_tilemap::tilemap;

let builder = TilemapBuilder::new();

// Equivalent to...

let builder = TilemapBuilder::default();

// Or...

let builder = Tilemap::builder();

pub fn topology(self, topology: GridTopology) -> TilemapBuilder

Sets the topology of the tilemap.

The default is a square grid. Use this if you want a hexagonal grid instead.

Examples

use bevy_tilemap::prelude::*;

let builder = TilemapBuilder::new().topology(GridTopology::HexY);

pub fn dimensions(self, width: u32, height: u32) -> TilemapBuilder

Sets the dimensions of the tilemap.

If this is not set then the tilemap will be boundless entirely.

Examples

use bevy_tilemap::prelude::*;

let builder = TilemapBuilder::new().dimensions(5, 5);

pub fn chunk_dimensions(
self,
width: u32,
height: u32,
depth: u32
) -> TilemapBuilder

Sets the chunk dimensions.

Chunk dimensions are in tiles. If this is not set then the default of 32x, 32y is used.

Examples

use bevy_tilemap::prelude::*;

let builder = TilemapBuilder::new().chunk_dimensions(32, 32, 1);

pub fn layer_offset(self, offset: Vec2) -> TilemapBuilder

Sets the layer offset as X, Y.

Per layer, this is the offset that will take place. This is generally useful for 3D style tilemaps that need to give the illusion of depth.

Examples

use bevy_tilemap::prelude::*;
use bevy_math::Vec2;

let builder = TilemapBuilder::new().layer_offset(Vec2::new(0.5, 0.5));

pub fn texture_dimensions(self, width: u32, height: u32) -> TilemapBuilder

Sets the tile dimensions.

Tile dimensions are in pixels. If this is not set then the default of 32px, 32px is used.

Examples

use bevy_tilemap::prelude::*;

let builder = TilemapBuilder::new().texture_dimensions(32, 32);

pub fn tile_scale(self, width: f32, height: f32, depth: f32) -> TilemapBuilder

Sets the tile scale.

By default this is the size of the texture for width and height which means that a tile width of 1 is the width in texture pixels for example.

Examples

use bevy_tilemap::prelude::*;

let builder = TilemapBuilder::new().tile_scale(32.0, 32.0, 32.0);

pub fn z_layers(self, layers: usize) -> TilemapBuilder

Sets the amount of render layers that sprites can exist on.

By default there are 20 if this is not set.

Examples

use bevy_tilemap::prelude::*;

let builder = TilemapBuilder::new().z_layers(5);

pub fn add_layer(
self,
layer: TilemapLayer,
sprite_order: usize
) -> TilemapBuilder

Adds a sprite layer that sprites can exist on.

Takes in a LayerKind and a Z layer and adds it to the builder.

If there are more layers than Z layers is set, builder will construct a tilemap with that many layers instead. In the case that a layer is added twice to the same Z layer, the first layer will be overwritten by the later.

Examples

use bevy_tilemap::prelude::*;

let builder = TilemapBuilder::new()
    .add_layer(TilemapLayer { kind: LayerKind::Dense, ..Default::default() }, 0)
    .add_layer(TilemapLayer { kind: LayerKind::Sparse, ..Default::default() }, 1)
    .add_layer(TilemapLayer { kind: LayerKind::Sparse, ..Default::default() }, 2);

pub fn texture_atlas(self, handle: Handle<TextureAtlas>) -> TilemapBuilder

Sets the texture atlas, this is required to be set.

Examples

use bevy_asset::{prelude::*, HandleId};
use bevy_sprite::prelude::*;
use bevy_tilemap::prelude::*;

let texture_atlas_handle = Handle::weak(HandleId::random::<TextureAtlas>());

let builder = TilemapBuilder::new().texture_atlas(texture_atlas_handle);

pub fn auto_chunk(self) -> Self

Sets if you want the tilemap to automatically spawn new chunks.

This is useful if the tilemap map is meant to be endless or nearly endless with a defined size. Otherwise, it probably is better to spawn chunks directly or creating a system that can automatically spawn and despawn them given context.

By default this is not enabled.

Examples

use bevy_tilemap::prelude::*;

let builder = TilemapBuilder::new().auto_chunk();

pub fn auto_spawn(self, width: u32, height: u32) -> Self

Sets the tilemap to automatically spawn new chunks within given dimensions.

This enables a feature which spawns just the right amount of chunks to fit the screen. It is possible that it may not be able to catch all dimensions but typical uses should be completely fine.

Examples

use bevy_tilemap::prelude::*;

let builder = TilemapBuilder::new().auto_spawn(2, 3);

pub fn finish(self) -> TilemapResult<Tilemap>

Consumes the builder and returns a result.

If successful a TilemapResult is return with tilemap on succes or a TilemapError if there is an issue.

Errors

If a texture atlas is not set this is the only way that an error can occur. If this happens, be sure to use texture_atlas.

Examples

use bevy_asset::{prelude::*, HandleId};
use bevy_sprite::prelude::*;
use bevy_tilemap::prelude::*;

let texture_atlas_handle = Handle::weak(HandleId::random::<TextureAtlas>());

let builder = TilemapBuilder::new().texture_dimensions(32, 32).texture_atlas(texture_atlas_handle);

assert!(builder.finish().is_ok());
assert!(TilemapBuilder::new().finish().is_err());